pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pop"
version = "0.1.0"
description = "Pop your code onto a VPS. Lightweight SSH-based deployment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paramiko>=3.0",
    "pyyaml>=6.0",
    "scp>=0.14",
    "rich>=13.0",
]

[project.scripts]
pop = "pop.cli:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-mock"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pop*"]
